backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.3
cloudinary==1.40.0
groq==1.1.2
hypothesis==6.108.5
pytest==8.2.2
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0